move mocks to json files on disk
