+++ tests/mocks/v2/last/trade/AAPL.json
@@ -1,0 +1,21 @@
+{
+    "results": {
+        "c": [
+            12,
+            37
+        ],
+        "i": "237688",
+        "p": 166.25,
+        "s": 2,
+        "x": 4,
+        "r": 202,
+        "z": 3,
+        "T": "AAPL",
+        "t": 1651179319310617300,
+        "y": 1651179319308000000,
+        "f": 1651179319310588400,
+        "q": 7084210
+    },
+    "status": "OK",
+    "request_id": "d4bafa50e72cf9ed19ac538ae1a3185a"
+}
