+++ tests/mocks/v2/snapshot/locale/global/markets/crypto/tickers/X:BTCUSD/book.json
@@ -1,0 +1,39 @@
+{
+    "data": {
+        "askCount": 593.1412981600005,
+        "asks": [
+            {
+                "p": 11454,
+                "x": {
+                    "2": 1
+                }
+            },
+            {
+                "p": 11455,
+                "x": {
+                    "2": 1
+                }
+            }
+        ],
+        "bidCount": 694.951789670001,
+        "bids": [
+            {
+                "p": 16303.17,
+                "x": {
+                    "1": 2
+                }
+            },
+            {
+                "p": 16302.94,
+                "x": {
+                    "1": 0.02859424,
+                    "6": 0.023455
+                }
+            }
+        ],
+        "spread": -4849.17,
+        "ticker": "X:BTCUSD",
+        "updated": 1605295074162
+    },
+    "status": "OK"
+}
