+++ tests/mocks/v2/snapshot/locale/us/markets/stocks/gainers?market.type=stocks.json
@@ -1,0 +1,100 @@
+{
+    "status": "OK",
+    "tickers": [
+        {
+            "day": {
+                "c": 6.42,
+                "h": 6.99,
+                "l": 6.4,
+                "o": 6.81,
+                "v": 115782,
+                "vw": 6.656
+            },
+            "lastQuote": {
+                "P": 6.43,
+                "S": 1,
+                "p": 6.4,
+                "s": 1,
+                "t": 1651251738312628478
+            },
+            "lastTrade": {
+                "c": [
+                    14,
+                    41
+                ],
+                "i": "100",
+                "p": 6.42,
+                "s": 200,
+                "t": 1651251334045891221,
+                "x": 8
+            },
+            "min": {
+                "av": 115689,
+                "c": 6.42,
+                "h": 6.542,
+                "l": 6.42,
+                "o": 6.49,
+                "v": 2671,
+                "vw": 6.4604
+            },
+            "prevDay": {
+                "c": 0.29,
+                "h": 0.348,
+                "l": 0.29,
+                "o": 0.3443,
+                "v": 1488660,
+                "vw": 0.317
+            },
+            "ticker": "NVCN",
+            "todaysChange": 6.13,
+            "todaysChangePerc": 2113.793,
+            "updated": 1651251360000000000
+        },
+        {
+            "day": {
+                "c": 4.2107,
+                "h": 4.95,
+                "l": 4.21,
+                "o": 4.31,
+                "v": 453199,
+                "vw": 4.4181
+            },
+            "lastQuote": {
+                "P": 4.22,
+                "S": 9,
+                "p": 4.21,
+                "s": 11,
+                "t": 1651251781709136903
+            },
+            "lastTrade": {
+                "c": null,
+                "i": "1084",
+                "p": 4.2116,
+                "s": 241,
+                "t": 1651251789345841015,
+                "x": 4
+            },
+            "min": {
+                "av": 453189,
+                "c": 4.2107,
+                "h": 4.2107,
+                "l": 4.2107,
+                "o": 4.2107,
+                "v": 1012,
+                "vw": 4.2107
+            },
+            "prevDay": {
+                "c": 0.1953,
+                "h": 0.2966,
+                "l": 0.195,
+                "o": 0.29,
+                "v": 8784033,
+                "vw": 0.2278
+            },
+            "ticker": "BIOL",
+            "todaysChange": 4.016,
+            "todaysChangePerc": 2056.477,
+            "updated": 1651251789345841015
+        }
+    ]
+}
