+++ tests/mocks/v3/reference/tickers/types.json
@@ -1,0 +1,97 @@
+{
+    "results": [
+        {
+            "code": "CS",
+            "description": "Common Stock",
+            "asset_class": "stocks",
+            "locale": "us"
+        },
+        {
+            "code": "PFD",
+            "description": "Preferred Stock",
+            "asset_class": "stocks",
+            "locale": "us"
+        },
+        {
+            "code": "WARRANT",
+            "description": "Warrant",
+            "asset_class": "stocks",
+            "locale": "us"
+        },
+        {
+            "code": "RIGHT",
+            "description": "Rights",
+            "asset_class": "stocks",
+            "locale": "us"
+        },
+        {
+            "code": "BOND",
+            "description": "Corporate Bond",
+            "asset_class": "stocks",
+            "locale": "us"
+        },
+        {
+            "code": "ETF",
+            "description": "Exchange Traded Fund",
+            "asset_class": "stocks",
+            "locale": "us"
+        },
+        {
+            "code": "ETN",
+            "description": "Exchange Traded Note",
+            "asset_class": "stocks",
+            "locale": "us"
+        },
+        {
+            "code": "SP",
+            "description": "Structured Product",
+            "asset_class": "stocks",
+            "locale": "us"
+        },
+        {
+            "code": "ADRC",
+            "description": "American Depository Receipt Common",
+            "asset_class": "stocks",
+            "locale": "us"
+        },
+        {
+            "code": "ADRW",
+            "description": "American Depository Receipt Warrants",
+            "asset_class": "stocks",
+            "locale": "us"
+        },
+        {
+            "code": "ADRR",
+            "description": "American Depository Receipt Rights",
+            "asset_class": "stocks",
+            "locale": "us"
+        },
+        {
+            "code": "FUND",
+            "description": "Fund",
+            "asset_class": "stocks",
+            "locale": "us"
+        },
+        {
+            "code": "BASKET",
+            "description": "Basket",
+            "asset_class": "stocks",
+            "locale": "us"
+        },
+        {
+            "code": "UNIT",
+            "description": "Unit",
+            "asset_class": "stocks",
+            "locale": "us"
+        },
+        {
+            "code": "LT",
+            "description": "Liquidating Trust",
+            "asset_class": "stocks",
+            "locale": "us"
+        }
+    ],
+    "status": "OK",
+    "request_id": "efbfc7c2304bba6c2f19a2567f568134",
+    "count": 15
+}
