+++ tests/mocks/v3/reference/tickers?cursor=YWN0aXZlPXRydWUmZGF0ZT0yMDIyLTA0LTI3JmxpbWl0PTImb3JkZXI9YXNjJnBhZ2VfbWFya2VyPUFBJTdDZjEyMmJjYmY4YWQwNzRmZmJlMTZmNjkxOWQ0ZDc3NjZlMzA3MWNmNmU1Nzg3OGE0OGU1NjQ1YzQyM2U3NzJhOSZzb3J0PXRpY2tlcg.json
@@ -1,0 +1,34 @@
+{
+    "results": [
+        {
+            "ticker": "AAA",
+            "name": "AAF First Priority CLO Bond ETF",
+            "market": "stocks",
+            "locale": "us",
+            "primary_exchange": "ARCX",
+            "type": "ETF",
+            "active": true,
+            "currency_name": "usd",
+            "composite_figi": "BBG00X5FSP48",
+            "share_class_figi": "BBG00X5FSPZ4",
+            "last_updated_utc": "2022-04-27T00:00:00Z"
+        },
+        {
+            "ticker": "AAAU",
+            "name": "Goldman Sachs Physical Gold ETF Shares",
+            "market": "stocks",
+            "locale": "us",
+            "primary_exchange": "BATS",
+            "type": "ETF",
+            "active": true,
+            "currency_name": "usd",
+            "cik": "0001708646",
+            "composite_figi": "BBG00LPXX872",
+            "share_class_figi": "BBG00LPXX8Z1",
+            "last_updated_utc": "2022-04-27T00:00:00Z"
+        }
+    ],
+    "status": "OK",
+    "request_id": "40d60d83fa0628503b4d13387b7bde2a",
+    "count": 2
+}
