+++ tests/mocks/v3/trades/AAPL?limit=2.json
@@ -1,0 +1,40 @@
+{
+    "results": [
+        {
+            "conditions": [
+                12,
+                37
+            ],
+            "correction": 1,
+            "exchange": 11,
+            "id": "183276",
+            "participant_timestamp": 1651181822461636600,
+            "price": 156.43,
+            "sequence_number": 7179341,
+            "sip_timestamp": 1651181822461979400,
+            "size": 10,
+            "tape": 3,
+            "trf_id": 3,
+            "trf_timestamp": 1651181557090806500
+        },
+        {
+            "conditions": [
+                12,
+                37
+            ],
+            "correction": 1,
+            "exchange": 12,
+            "id": "183276",
+            "participant_timestamp": 1651181822461636600,
+            "price": 157.43,
+            "sequence_number": 7179341,
+            "sip_timestamp": 1651181822461979400,
+            "size": 10,
+            "tape": 3,
+            "trf_id": 3,
+            "trf_timestamp": 1651181557090806500
+        }
+    ],
+    "status": "OK",
+    "request_id": "756f9910624b35a47eb07f21a7a373bb"
+}
