+++ tests/test_aggs.py
@@ -1,4 +1,4 @@
-from mocks import BaseTest
+from base import BaseTest
 from polygon.rest.models import (
     Agg,
     GroupedDailyAgg,
